separated two modules data_extraction.py and magnetism.py from workflows.py
